requirements.txt
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
idna==2.8
pycodestyle==2.5.0
python-dotenv==0.10.1
requests==2.21.0
schedule==0.6.0
urllib3==1.24.1
